+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## {Next Version}
+## {{Next Version}}
 
 ### Bugfixes
 
-- Don't let reveal clobber secret files (#579)
+- In 'tell', warn about disabled, revoked, expired, or invalid keys (#552, #508, #317, #290, #283, #238)
+- Error if 'tell' is used on an email address with multiple keys (#552)
+- Don't let 'reveal' clobber secret files (#579)
 
 ### Misc
 
@@ -34,7 +36,7 @@
 ### Features
 
 - Support SECRETS_PINENTRY env var for gnupg --pinentry-mode parameter (#221)
-- Show output from gnupg if 'hide' fails (#516)
+- Show output from gnupg if 'hide' fails (#516, #202, #317)
 - Add support for Busybox (#478)
 
 ### Bugfixes